+++ webcomponents/math/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.0.1 (2020-10-17)
+
+### Fix
+
+- try to fix publish to unpkg
+
 # 2.0.0 (2020-09-03)
 
 ### Breaking Changes